UI: Display invalid cache as striped lines in the timeline
Unify the way the different state's of a cache are shown in the
timeline:

* Baked: fully opaque
* Cached: slightly transparent
* Invalid cache: slightly transparent, dark diagonal stripes

This improves accessibility since patterns are easier to recognize
for colorblind or otherwise visually impaired people.

The slight transparency is done with an alpha of 0.7 and the diagonal
stripes use the cache's color at 50% value.

Implements https://projects.blender.org/blender/blender/issues/108196