Avoid infinite update loops caused by syncing of flags.
Modifiers store "expansion" flags in addition to the panels' own
open/closed states. This is so panel state is persistent when editors
are switched, which destroys and recreates the panels.

The expansion flags are applied to panels twice: An initial update
after creating a panel makes sure the flag matches panel state, a
general update happens in `UI_panels_end` to update existing panels.

The syncing function for panel state was assuming that instancing would
only happen for top-level panels. Now that child panels can also use
instancing the flag index to sync needs to be computed from the child's
position in the list. It's important that the index for gettings and
setting the panel state is correct, otherwise a wrong panel is used and
the expected update does not happen. That can lead to infinite drawing
updates because the expected change never happens and the syncing
function keeps tagging the region for redraw.