Assets: add function to copy asset data to another ID
Add a function `some_id.asset_data.copy_to_id(other_id)` to copy the asset
data from one data-block to another. This is intended to be used in the
pose library, when updating a pose by simply creating a new asset and
having that replace the old one.

This is intentionally taking a copy, even though the above use case could
have sufficed with a higher-level 'move' function. By exposing this as a
copy, it can be used in a wider range of situations, from whatever Python
code wants to use it. This could include copying the asset data from the
active asset to all the other selected ones.

Any pre-existing asset data is freed before the copy is assigned. The
target ID doesn't have to be marked as asset yet for this function to work.
Note that using this function will make the destination ID an asset,
without going through the regular operator.