Allow forcing viewport view layer depsgraph to use rendered mode
This is a debugging tool for technical artists, to help nailing down
render artifacts while having all troubleshooting tools available:
for example, easier navigation, objects visibility control, geometry
nodes spreadsheet, etc.

Available as a "mode" changing button in a new Dependency Graph panel
in the View Layer buttons. The evaluation mode is not stored in the
.blend file to avoid creation of possible confusion.