Fix #108578: Crash when unlinking multi-input sockets
Fix a mistake in commit 2ce5fc4a3e that caused a crash when detaching
node links from multi-input sockets.

When a link is detached `nodeRemLink` nulls the `link` pointer of the
socket. So before the next update multi inputs are linked but don't
have a valid `link` pointer causing the crash, when trying to access
the link in `std_node_socket_draw`.

The introduced check avoids the crash and is more correct since it
doesn't just check one link for multi-input sockets.