UI: move 'Face Nearest' snap option to another session
The snap mode called "Face Nearest" (and the "Increment" but that's for
another time) doesn't behave like the other snap modes.

Unlike the other snap modes, "Face Nearest" does not act on a Snap
Source (also called Snap Base or Snap With). It always acts on the
origin of individually transformed elements, such as each vertex
individually.

It works just like the "Project Individual Elements" option. Therefore,
it would be more convenient for it to be an option like "Project
Individual Elements".