Fix #106395: Custom properties dont appear in quick favourites
Issue here was that Quick Favorites use a property's **identifier** when
adding [which is the bare name without brackets etc. from id properties]
but when spawning the actual menu,  `RNA_struct_find_property` expects
the identifier to already include the brackets to know these are id
properties (later on in `screen_user_menu_draw`).

So to solve this, now include the needed syntax when storing the
`bUserMenuItem_Prop` identfier.
Seems the quickest way to append the needed characters is using
`RNA_path_property_py` (not sure if there a re better ways to do this).
Also note that we (need to) ignore the actual array index constructing
the string [alays pass -1 here] since the index is handled separately [I
tested boolean arrays and these work].