DEV: Propagate errors in make.bat
Chained commands can be executed from the command line, for example, we
can execute a command to compile and we can chain another command to
execute the program that we are compiling as soon as it is ready.

However, when running make.bat, if an error occurs, it just terminates
without returning error, if we want to compile and run blender only if
it was a successful build, since make.bat never propagates errors,
we will run an outdated version of blender.

With the patch the error is propagated in order to be able to chain
commands with make.bat.