Fix compilation error with oneAPI disabled and GPU Embree enabled
Was only happening on the Gentoo system used at the studio: the linking state
was failing with some missing symbols and print about missing libsycl.so.6.

The reason for this is not really clear to me: to my knowledge we do not use
libsycl when oneAPI device is disabled, and we should not be passing the
indirect dependencies of shared libraries to the linker. It is also something
that I was unable to reproduce on Debian 12 system.

Nevertheless, the current state of the CMakeLists.txt does request the
cycles_bvh to be linked against SYCL_LIBRARIES. However, the SYCL was only
requested to be found if WITH_CYCLES_DEVICE_ONEAP is true.

This change makes it so the SYCL is requested to be found if the oneAPI
device is enabled, or if the Embree is detected to require/use SYCL
support.