Fix #108911: Fcurve Proportional editing circle draws wrong
Since we set up the GPU matrix a certain way (in order to draw a round
circle even for different UI x/y scaling), we have to tweak where the
center of that circle draws accordingly (instead of passing the raw
frame/value as coords).

There might be ways to solve this differently (e.g. setting up matrix
manipulation different or having a version of imm_drawcircball that can
do this in a way that compensates UI view scaling), but havent found a
quick way to do this and fixing it like done here seemed straightforward
enough to make a PR from.

Should probably go into 3.3 LTS as well.