Initial work to manage preview via asset representation & load via UI
Basic work to let the asset representation hold the preview image, which
can be requested by the UI. The asset/file browser still requests the
preview, but the UI does the actual loading now. This means we don't
depend on the fragile file preview caching anymore.

There are still some issues to be fixed.