deps_builder: output html report for release manager
request from Thomas, to update the licensing document he needs
a list of all dependencies their versions and their homepage.

for all deps hosted on github the script will do its best to
figure out the landing page, for all others DEPNAME_HOMEPAGE
will have to be set in versions.cmake

To ensure this information is always supplied, the script will
error out with a fatal error informing whoever is working on
the builder to supply this information.