Tracking: Add control over margin between pattern and search
Measured in percentage relative to the search area size, and is
ensured on a motion tracking step.

The margin is always ensured upon a successful tracking step,
which might lead to big search area when affine/perspective
tracker decides to squish the marker. Proper solution for this
problem seems to add an extra constraint and disallow tracker
from returning long stretched marker, as it will not be very
useful situation.

Ref #107487