Transform Snap: Deduplicate code and test clip planes in boundbox
This commit deduplicates the code of:
- Create local projection matrix;
- Create local Clip Planes;
- Test snap in BoundBox;

These common codes are repeated for each type of object (Empty, Camera, Armature, Curve, Mesh, Editmesh...).

Now they have been unified in the creation of `Nearest2dUserData`.

Changes:
- Local Clip Panes are created in the `Nearest2dUserData` constructor.
So these Clip Planes can now be used to test the BoundBox's occlusion and return early. The downside is that it adds a bit more code and can be a bit bad for performance.