Python image_utils: remove unnecessary os.path.exists calls
I recently ran into a performance issue in FBX import where the imported
file referenced many non-existent images, but with filepaths with the
same drive letter as a DVD drive. The FBX importer would search for the
images with image_utils.load_image, but this would take up the majority
of the total import time due to the 100 times slower IO of the DVD drive
compared to an HDD or SSD.

Filepath variants can end up the same, so this patch will skip the
duplicates. This is the main performance gain in my specific FBX import
case because it removes half of the checks for filepaths existing on the
DVD drive.

bpy.path.resolve_ncase already checks whether `filepath_test` exists, so
there was no need to separately check whether it exists in the
`ncase_cmp` case that uses bpy.path.resolve_ncase. Because there's no
simple way to tell whether bpy.path.resolve_ncase found a valid path or
returned its `path` argument, a new `default_orig=True` argument has
been added which, when set to False, will cause the function to return
None when the path is not found. Now, with `default_orig=False`, if the
returned value is not None, it is known that the returned path does
exist, so no additional checks are required.

There are no functional changes expected with this patch aside from
exceptionally rare edge-cases where, because the original code would
check `filepath_test` multiple times, `filepath_test` could not exist
the first time it was checked, but then would exist by the next time it
was checked.