Use a single list of "items" that can be either sockets or panels.
This makes it possible to arrange sockets and panels more flexibly.
If sockets and panels are in separate lists there has to be a rule that
says which kind comes first in the UI. If we want to be able to move
sockets in front of panels and vice versa having a common list is
necessary.