Cycles: oneAPI: Improve performance of scenes not using volume
The NODE_VOLUME feature adds a lot of code to shade_surface,
including the entire NanoVDB library, causing GPU compilers to
spend additional registers while the feature is not used often.
This leads to suboptimal execution and the use of a specialized
version gives a noticeable speed-up for oneAPI devices.