Fix #109236: Split Edges node skips loose edges
With the previous fix to the node, 8a11f0f3a23ef9b3c11c, the new edge
indices are build with the changed corner verts from a previous step in
the algorithm. That doesn't work for loose edges though, since they
aren't used by any face corners. The best solution I could come up with
was adding a second loop over the split vertices that adjusts the vertex
indices of loose edges. This can be skippe when there are none.