Simplified base color for socket types
Socket types have a `draw_color` callback that defines the color of a
socket in the ui (node editor, group interfaces, shader tree view, etc).
This callback is overly generic, it takes a `bContext` as well as a node
and socket instance. Technically this callback can only be used when
there is a context and a concrete instance of a socket.

In the vast majority of cases this contextual data is not necessary and
we simply use a static type-based color. To make this less complicated,
this patch adds an alternative `base_color` type member. This is a plain
color value which is used in place of the `draw_color` callback when it
isn't defined. Standard socket types no longer define a `draw_color`
callback and only use the `base_color`.

Compatibility with python nodes addons is important:
- The `draw_color` callback was required, it is now optional. Addons can
  keep using it, but `bl_base_color` is recommended for new code.
- The new `base_color` property is registerable, but optional so as to
  not cause errors in existing scripts.
- [ ] Documentation needs updating:
  - Explain that `draw_color` is now optional, and `bl_base_color` is
    used by default.
  - Suggest defining the `bl_base_color` property instead of the
    `draw_color` function.
- [ ] Update python templates for custom nodes to demonstrate the
  `bl_base_color` property.