Fix resolve_ncase not searching the current working directory when given a relative path without a leading "./"
Given a file "test.txt" existing in a case-sensitive current working
directory, resolve_ncase would not find the file given the relative path
"TEST.txt" because the `dirpath` would be an empty string which is not a
valid path to the current working directory.

This patch instead searches the current working directory whenever
`dirpath` is the empty string.

There is no need to check whether the current working director is a path
that exists or whether it is a directory, so these checks can be
skipped. In doing so, recursive calls to _ncase_path_found can never
have the empty string as their `path` argument, so checking for the
empty string at the start of _ncase_path_found has been moved out to the
start of resolve_ncase.