Make sequencer_preview_add_sound retry if the PreviewJob is terminating
If a sound is added to the queue when the preview job is done processing
its current list of sounds, a leak could happen: The new sound gets
added to the queue but the job never actually processes it.

This change follows @iss's suggestion of using the WM_event_add_notifier
function to force a redraw when that is that case.

If the PreviewJob is terminating (running = false):
- First clear the sound's loading tag: Without this, the code in
  sequencer_draw.c will not be able to push the sound for processing in
  the second pass.
- Add the event notifier for the current scene.