Re-update Eigen to the version 3.3.7
The previous update in the b19c437eff7 went quite messy:
basically due to too much trust put to the legacy update
script which does not take care of files which were removed
from the upstream and not applying local patches.

Updated the script to solve the problem of stale files by
removing the bundled directory. Also removed the patch
directory as the patch in there was not applied in the past
4 years, without any reports.