Fix #109433: Snapping occluding objects even in wireframe mode
The internal `use_occlusion_test` option was only being removed if it was in conjunction with the `Snap to Face` option.

Both test occlusion and Snap to Face are conflicting. In wireframe mode the objects are not occluded, but Snap to Face can prevent them from being snapped.

The proposed solution is to prioritize snapping to other non-Face elements but still allow "Snap to Face" in X-Ray mode.